backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.18

# Data Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# NLP & ML
spacy==3.7.5
stanza==1.8.2
torch==2.6.0
numpy==1.26.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
pdf2image==1.17.0
Pillow==10.4.0
PyMuPDF>=1.23.0  # For PDF overlay anonymization
pytesseract>=0.3.10  # Optional: OCR for images

# Utilities
python-dateutil==2.9.0
flashtext==2.7

# Server
gunicorn==22.0.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
black==24.8.0
